ATR-917: updated unit tests for PX1054, added check for IsActive and IsActiveFor
